split sga archive classes into core and graph
