+++ AOEMods.Essence.Editor/ArchivePropertiesViewModel.cs
@@ -1,4 +1,4 @@
-﻿using AOEMods.Essence.SGA;
+﻿using AOEMods.Essence.SGA.Graph;
 using Microsoft.Toolkit.Mvvm.ComponentModel;
 using Microsoft.Toolkit.Mvvm.Input;
 using System;
