+++ AOEMods.Essence.Editor/ExportArchiveUtil.cs
@@ -3,6 +3,7 @@ using AOEMods.Essence.Chunky.RGD;
 using AOEMods.Essence.Chunky.RRMaterial;
 using AOEMods.Essence.Chunky.RRTex;
 using AOEMods.Essence.SGA;
+using AOEMods.Essence.SGA.Graph;
 using Microsoft.Toolkit.Mvvm.Messaging;
 using Ookii.Dialogs.Wpf;
 using SharpGLTF.Materials;
