--- AOEMods.Essence/SGA/ArchiveFolderEntry.cs
@@ -1,3 +1,0 @@
-﻿namespace AOEMods.Essence.SGA;
-
-public record ArchiveFolderEntry(uint NameOffset, uint FolderStartIndex, uint FolderEndIndex, uint FileStartIndex, uint FileEndIndex) : IArchiveEntry;
