--- AOEMods.Essence/SGA/ArchiveTocEntry.cs
@@ -1,3 +1,0 @@
-﻿namespace AOEMods.Essence.SGA;
-
-public record ArchiveTocEntry(string Alias, string Name, uint FolderStartIndex, uint FolderEndIndex, uint FileStartIndex, uint FileEndIndex, uint FolderRootIndex) : IArchiveEntry;
