+++ AOEMods.Essence/SGA/Core/ArchiveEntries.cs
@@ -1,3 +1,3 @@
-﻿namespace AOEMods.Essence.SGA;
+﻿namespace AOEMods.Essence.SGA.Core;
 
 public record ArchiveEntries(ArchiveHeader Header, IList<ArchiveTocEntry> Tocs, IList<ArchiveFolderEntry> Folders, IList<ArchiveFileEntry> Files);
