+++ AOEMods.Essence/SGA/Core/ArchiveFolderEntry.cs
@@ -1,0 +1,3 @@
+﻿namespace AOEMods.Essence.SGA.Core;
+
+public record ArchiveFolderEntry(uint NameOffset, uint FolderStartIndex, uint FolderEndIndex, uint FileStartIndex, uint FileEndIndex);
