+++ AOEMods.Essence/SGA/Core/ArchiveHeader.cs
@@ -1,4 +1,4 @@
-﻿namespace AOEMods.Essence.SGA;
+﻿namespace AOEMods.Essence.SGA.Core;
 
 public record ArchiveHeader(
     byte[] Magic,
