+++ AOEMods.Essence/SGA/Core/ArchiveTocEntry.cs
@@ -1,0 +1,3 @@
+﻿namespace AOEMods.Essence.SGA.Core;
+
+public record ArchiveTocEntry(string Alias, string Name, uint FolderStartIndex, uint FolderEndIndex, uint FileStartIndex, uint FileEndIndex, uint FolderRootIndex);
