+++ AOEMods.Essence/SGA/Core/FileStorageType.cs
@@ -1,4 +1,4 @@
-namespace AOEMods.Essence.SGA;
+namespace AOEMods.Essence.SGA.Core;
 
 public enum FileStorageType : byte
 {
