+++ AOEMods.Essence/SGA/Graph/ArchiveFolderNode.cs
@@ -1,4 +1,4 @@
-﻿namespace AOEMods.Essence.SGA;
+﻿namespace AOEMods.Essence.SGA.Graph;
 
 public class ArchiveFolderNode : IArchiveFolderNode
 {
