+++ AOEMods.Essence/SGA/Graph/ArchiveMemoryFileNode.cs
@@ -1,4 +1,4 @@
-﻿namespace AOEMods.Essence.SGA;
+﻿namespace AOEMods.Essence.SGA.Graph;
 
 public class ArchiveMemoryFileNode : IArchiveFileNode
 {
