+++ AOEMods.Essence/SGA/Graph/ArchiveStoredFileNode.cs
@@ -1,4 +1,4 @@
-﻿namespace AOEMods.Essence.SGA;
+﻿namespace AOEMods.Essence.SGA.Graph;
 
 public class ArchiveStoredFileNode : IArchiveFileNode
 {
