+++ AOEMods.Essence/SGA/Graph/ArchiveToc.cs
@@ -1,4 +1,4 @@
-﻿namespace AOEMods.Essence.SGA;
+﻿namespace AOEMods.Essence.SGA.Graph;
 
 public class ArchiveToc : IArchiveToc
 {
