+++ AOEMods.Essence/SGA/Graph/IArchiveFileNode.cs
@@ -1,4 +1,4 @@
-﻿namespace AOEMods.Essence.SGA
+﻿namespace AOEMods.Essence.SGA.Graph
 {
     public interface IArchiveFileNode : IArchiveNode
     {
