+++ AOEMods.Essence/SGA/Graph/IArchiveNode.cs
@@ -1,4 +1,4 @@
-﻿namespace AOEMods.Essence.SGA;
+﻿namespace AOEMods.Essence.SGA.Graph;
 
 public interface IArchiveNode
 {
