--- AOEMods.Essence/SGA/IArchiveEntry.cs
@@ -1,5 +1,0 @@
-﻿namespace AOEMods.Essence.SGA;
-
-public interface IArchiveEntry
-{
-}
